+++ mllib/src/main/scala/org/apache/spark/ml/regression/GBTRegressor.scala
@@ -228,14 +228,14 @@ final class GBTRegressionModel private[ml](
    * Estimate of the importance of each feature.
    *
    * Each feature's importance is the average of its importance across all trees in the ensemble
-   * The importance vector is normalized to sum to 1. This method is suggested by Friedman
-   * (J. H. Friedman, "Greedy Function Approximation: A Gradient Boosting Machine", 1999)
+   * The importance vector is normalized to sum to 1. This method is suggested by Hastie et al.
+   * (Hastie, Tibshirani, Friedman. "The Elements of Statistical Learning, 2nd Edition." 2001.)
    * and follows the implementation from scikit-learn.
    *
    * @see [[DecisionTreeRegressionModel.featureImportances]]
    */
   @Since("2.0.0")
-  lazy val featureImportances: Vector = featureImportances(numFeatures)
+  lazy val featureImportances: Vector = TreeEnsembleModel.featureImportances(trees, numFeatures)
 
   /** (private[ml]) Convert to a model in the old API */
   private[ml] def toOld: OldGBTModel = {
